try to symlink REAMDE

+++ run_website.py
@@ -20,6 +20,8 @@ def kill_port_unix(port):
         result = subprocess.check_output(command, shell=True).decode("utf-8")
         pids = result.strip().split("\n")
         for pid in pids:
+			if (pid == "0"):
+				continue
             subprocess.run(f"kill -9 {pid}", shell=True)
     except subprocess.CalledProcessError as e:
         return
@@ -30,6 +32,8 @@ def kill_port_windows(port):
         result = subprocess.check_output(command, shell=True).decode("utf-8").strip().split("\n")
         for line in result:
             pid = line.split()[-1]
+			if (pid == "0"):
+				continue
             subprocess.run(f"taskkill /F /PID {pid}", shell=True)
     except subprocess.CalledProcessError as e:
         return
@@ -57,7 +61,7 @@ class Handler(http.server.SimpleHTTPRequestHandler):
 def start_server(port):
     try:
         with socketserver.ThreadingTCPServer(("", port), Handler) as httpd:
-            print(f'Serving at port {port}. Visit http://localhost:{port}. To stop, press Ctrl+C (multiple times if needed)')
+            print(f'Visit http://localhost:{port} for ViralWasm-Consensus. Please leave this window open! To stop, press Ctrl+C (multiple times if needed)')
             webbrowser.open(f"http://localhost:{port}")
             httpd.serve_forever()
     except KeyboardInterrupt:
